+++ src/components/no-movies.js
@@ -1,4 +1,4 @@
-import {createElement} from '../util';
+import {createElement} from '../utils/util';
 
 const createNoMovieTemplate = () =>
   `<h2 class="films-list__title">There are no movies in our database</h2>`;
