+++ src/utils/render.js
@@ -1,0 +1,17 @@
+import {renderPosition} from "./util";
+
+export const render = (container, element, place = renderPosition.BEFOREEND) => {
+  switch (place) {
+    case renderPosition.BEFOREEND:
+      container.append(element);
+      break;
+    case renderPosition.AFTEREND:
+      container.insertAdjacentElement(renderPosition.AFTEREND, element);
+      break;
+  }
+};
+
+export const remove = (component) => {
+  component.getElement().remove();
+  component.removeElement();
+};
